Рефакторинг: - переписаны все сервисы - вся обработка сетевых запросов перенесена в сетевой клиент - сервисы больше не содержат в себе инициализаторы (обязательные) - сервисы готовы к тестам
