+++ TravelSchedule/Helper/Config/Constants.swift
@@ -1,5 +1,8 @@
 import Foundation
+import OpenAPIRuntime
+import OpenAPIURLSession
 
 struct Constants {
     static let apiKey: String = "9995803a-ea85-45dd-8f60-d8ae279985d8"
 }
+
