+++ TravelSchedule/Services/SearchService.swift
@@ -1,39 +1,24 @@
 import OpenAPIRuntime
 import OpenAPIURLSession
-import Foundation
 
-
+// MARK: - Protocol
 protocol SearchServiceProtocol {
     func search(from: String, to: String, transfers: Bool) async throws -> SegmentsResponse
 }
 
+// MARK: - Service
 final class SearchService: SearchServiceProtocol {
-    private let apiKey: String
-    private let client: Client
+    private let networkClient: NetworkClient
     
-    init(apiKey: String, client: Client) {
-        self.apiKey = apiKey
-        self.client = client
+    init(networkClient: NetworkClient = NetworkClient()) {
+        self.networkClient = networkClient
     }
     
-    func search(from: String, to: String, transfers: Bool = true) async throws -> SegmentsResponse {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        let todayString = dateFormatter.string(from: Date())
-        
-        do {
-            let response = try await client.getSchedualBetweenStations(query: .init(
-                apikey: apiKey,
-                from: from,
-                to: to,
-                date: todayString,
-                transfers: transfers
-            ))
-            return try response.ok.body.json
-        } catch URLError.Code.notConnectedToInternet {
-            throw NetworkError.noInternet
-        } catch {
-            throw NetworkError.serverError
-        }
+    func search(
+        from: String,
+        to: String,
+        transfers: Bool = true
+    ) async throws -> SegmentsResponse {
+        return try await networkClient.search(from: from, to: to, transfers: transfers)
     }
 }
